pyproject.toml
[project]
name = "charged"
version = "0.1.0"
description = "Hack4Rail Challenge 9 - Monitoring and management of railway vehicle batteries."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "nbstripout>=0.8.1",
    "plotly>=6.1.2",
    "pre-commit>=4.2.0",
    "pre-commit-hooks>=5.0.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
    "snowflake-snowpark-python[pandas]>=1.33.0",
]

[tool.hatch.build.targets.sdist]
include = ["src/charged"]

[tool.hatch.build.targets.wheel]
include = ["src/charged"]

[tool.hatch.build.targets.wheel.sources]
"src/charged" = "charged"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"